app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
httpx
pytest